+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
- * Add a snackbar add a Snackbar showing when the application has been updated.
+ * Add a snackbar add a Snackbar showing when the application has been updated. and if the underlying Graphviz version has been updated or not.
  * Make the version in the about dialog a link to the version in CHANGELOG.md.
  * Bundle @hpcc-js/wasm instead of loading from unpkg
  * Upgraded d3-graphviz to version 3.0.6, thereby upgrading @hpcc-js-/wasm to version 0.3.13, thereby upgrading Graphviz to version 2.44.0.